+++ vtkAddAdditionalAttribute.h
@@ -60,7 +60,8 @@ private:
 	// Reads this file in as an additional attribute array.
  //  If a marked file was specified, it reads in 
 	// only the particles at indices which were marked
-	int ReadAdditionalAttributeFile(vtkstd::vector<int>& markedParticleIndices,
+	int ReadAdditionalAttributeFile(
+		vtkstd::vector<unsigned long>& markedParticleIndices, 
 		vtkPointSet* output);
 //ETX
 };
